Moved and refactored organization entities
