+++ frontend/src/core/models/organization/new-organization-model.ts
@@ -1,0 +1,4 @@
+export interface NewOrganizationModel {
+    name: string,
+    authorId: number
+}
