+++ frontend/src/core/models/organization/organization-model.ts
@@ -1,4 +1,4 @@
-﻿export interface OrganizationModel{
+﻿export interface OrganizationModel {
   id: number,
   name: string,
   authorId: number,
