--- frontend/src/entity-models/organization-model.ts
@@ -1,6 +1,0 @@
-export interface OrganizationModel{
-    id?: number,
-    name: string,
-    authorId: number
-}
-
